src/analyst/requirements.txt
boto3==1.34.153
botocore==1.34.153
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
